[ROSTESTS][CLANG] Fix test build warnings
Problem: Clang builds of the ReactOS and Wine test suites report format, type, qualifier, and compatibility warnings, with some configurations treating them as build failures.

Why: Test diagnostics and helper macros retained mismatched argument types, while several test sources relied on compiler-specific declarations or warning-prone equivalent expressions.

How: Match diagnostic formats and helper casts to their exact types, add compile-only compatibility declarations and qualifiers, and rewrite equivalent expressions without changing test expectations or exercised behavior.

+++ modules/rostests/winetests/userenv/userenv.c
@@ -31,9 +31,9 @@
 
 #include "wine/test.h"
 
-#define expect(EXPECTED,GOT) ok((GOT)==(EXPECTED), "Expected %d, got %d\n", (EXPECTED), (GOT))
-#define expect_env(EXPECTED,GOT,VAR) ok((GOT)==(EXPECTED), "Expected %d, got %d for %s (%d)\n", (EXPECTED), (GOT), (VAR), j)
-#define expect_gle(EXPECTED) ok(GetLastError() == (EXPECTED), "Expected %d, got %d\n", (EXPECTED), GetLastError())
+#define expect(EXPECTED,GOT) ok((GOT)==(EXPECTED), "Expected %ld, got %ld\n", (LONG)(EXPECTED), (LONG)(GOT))
+#define expect_env(EXPECTED,GOT,VAR) ok((GOT)==(EXPECTED), "Expected %ld, got %ld for %s (%d)\n", (LONG)(EXPECTED), (LONG)(GOT), (VAR), j)
+#define expect_gle(EXPECTED) ok(GetLastError() == (EXPECTED), "Expected %lu, got %lu\n", (DWORD)(EXPECTED), GetLastError())
 
 static BOOL (WINAPI *pIsWow64Process)(HANDLE,PBOOL);
 
@@ -112,7 +112,7 @@ static void test_create_env(void)
     expect(TRUE, r);
 
     r = GetEnvironmentVariableA("SystemRoot", systemroot, sizeof(systemroot));
-    ok(r != 0, "GetEnvironmentVariable failed (%d)\n", GetLastError());
+    ok(r != 0, "GetEnvironmentVariable failed (%lu)\n", GetLastError());
 
     r = SetEnvironmentVariableA("SystemRoot", "overwrite");
     expect(TRUE, r);
@@ -237,7 +237,7 @@ static void test_get_profiles_dir(void)
     char *profiles_dir, *buf, small_buf[1];
 
     l = RegOpenKeyExA(HKEY_LOCAL_MACHINE, ProfileListA, 0, KEY_READ, &key);
-    ok(!l, "RegOpenKeyExA failed: %d\n", GetLastError());
+    ok(!l, "RegOpenKeyExA failed: %lu\n", GetLastError());
 
     l = RegQueryValueExA(key, ProfilesDirectory, NULL, NULL, NULL, &cch);
     if (l)
@@ -269,7 +269,7 @@ static void test_get_profiles_dir(void)
     /* MSDN claims the returned character count includes the NULL terminator
      * when the buffer is too small, but that's not in fact what gets returned.
      */
-    ok(cch == profiles_len - 1, "expected %d, got %d\n", profiles_len - 1, cch);
+    ok(cch == profiles_len - 1, "expected %lu, got %lu\n", profiles_len - 1, cch);
     /* Allocate one more character than the return value to prevent a buffer
      * overrun.
      */
@@ -317,26 +317,26 @@ static void test_get_user_profile_dir(void)
     }
 
     ret = OpenProcessToken( GetCurrentProcess(), TOKEN_QUERY, &token );
-    ok(ret, "expected success %u\n", GetLastError());
+    ok(ret, "expected success %lu\n", GetLastError());
 
     SetLastError( 0xdeadbeef );
     ret = GetUserProfileDirectoryA( NULL, NULL, NULL );
     error = GetLastError();
     ok(!ret, "expected failure\n");
-    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %u\n", error);
+    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %lu\n", error);
 
     SetLastError( 0xdeadbeef );
     ret = GetUserProfileDirectoryA( token, NULL, NULL );
     error = GetLastError();
     ok(!ret, "expected failure\n");
-    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %u\n", error);
+    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %lu\n", error);
 
     dirA = HeapAlloc( GetProcessHeap(), 0, 32 );
     SetLastError( 0xdeadbeef );
     ret = GetUserProfileDirectoryA( token, dirA, NULL );
     error = GetLastError();
     ok(!ret, "expected failure\n");
-    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %u\n", error);
+    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %lu\n", error);
     HeapFree( GetProcessHeap(), 0, dirA );
 
     len = 0;
@@ -344,8 +344,8 @@ static void test_get_user_profile_dir(void)
     ret = GetUserProfileDirectoryA( token, NULL, &len );
     error = GetLastError();
     ok(!ret, "expected failure\n");
-    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %u\n", error);
-    ok(!len, "expected 0, got %u\n", len);
+    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %lu\n", error);
+    ok(!len, "expected 0, got %lu\n", len);
 
     len = 0;
     dirA = HeapAlloc( GetProcessHeap(), HEAP_ZERO_MEMORY, 32 );
@@ -353,16 +353,16 @@ static void test_get_user_profile_dir(void)
     ret = GetUserProfileDirectoryA( token, dirA, &len );
     error = GetLastError();
     ok(!ret, "expected failure\n");
-    ok(error == ERROR_INSUFFICIENT_BUFFER, "expected ERROR_INSUFFICIENT_BUFFER, got %u\n", error);
+    ok(error == ERROR_INSUFFICIENT_BUFFER, "expected ERROR_INSUFFICIENT_BUFFER, got %lu\n", error);
     ok(len, "expected len > 0\n");
     HeapFree( GetProcessHeap(), 0, dirA );
 
     dirA = HeapAlloc( GetProcessHeap(), HEAP_ZERO_MEMORY, len );
     SetLastError( 0xdeadbeef );
     ret = GetUserProfileDirectoryA( token, dirA, &len );
-    ok(ret, "expected success %u\n", GetLastError());
+    ok(ret, "expected success %lu\n", GetLastError());
     ok(len, "expected len > 0\n");
-    ok(lstrlenA( dirA ) == len - 1, "length mismatch %d != %d - 1\n", lstrlenA( dirA ), len );
+    ok(lstrlenA( dirA ) == len - 1, "length mismatch %d != %lu - 1\n", lstrlenA( dirA ), len );
     trace("%s\n", dirA);
     HeapFree( GetProcessHeap(), 0, dirA );
 
@@ -370,20 +370,20 @@ static void test_get_user_profile_dir(void)
     ret = GetUserProfileDirectoryW( NULL, NULL, NULL );
     error = GetLastError();
     ok(!ret, "expected failure\n");
-    todo_wine ok(error == ERROR_INVALID_HANDLE, "expected ERROR_INVALID_HANDLE, got %u\n", error);
+    todo_wine ok(error == ERROR_INVALID_HANDLE, "expected ERROR_INVALID_HANDLE, got %lu\n", error);
 
     SetLastError( 0xdeadbeef );
     ret = GetUserProfileDirectoryW( token, NULL, NULL );
     error = GetLastError();
     ok(!ret, "expected failure\n");
-    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %u\n", error);
+    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %lu\n", error);
 
     dirW = HeapAlloc( GetProcessHeap(), 0, 32 );
     SetLastError( 0xdeadbeef );
     ret = GetUserProfileDirectoryW( token, dirW, NULL );
     error = GetLastError();
     ok(!ret, "expected failure\n");
-    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %u\n", error);
+    ok(error == ERROR_INVALID_PARAMETER, "expected ERROR_INVALID_PARAMETER, got %lu\n", error);
     HeapFree( GetProcessHeap(), 0, dirW );
 
     len = 0;
@@ -391,15 +391,15 @@ static void test_get_user_profile_dir(void)
     ret = GetUserProfileDirectoryW( token, NULL, &len );
     error = GetLastError();
     ok(!ret, "expected failure\n");
-    ok(error == ERROR_INSUFFICIENT_BUFFER, "expected ERROR_INSUFFICIENT_BUFFER, got %u\n", error);
+    ok(error == ERROR_INSUFFICIENT_BUFFER, "expected ERROR_INSUFFICIENT_BUFFER, got %lu\n", error);
     ok(len, "expected len > 0\n");
 
     dirW = HeapAlloc( GetProcessHeap(), HEAP_ZERO_MEMORY, len * sizeof(WCHAR) );
     SetLastError( 0xdeadbeef );
     ret = GetUserProfileDirectoryW( token, dirW, &len );
-    ok(ret, "expected success %u\n", GetLastError());
+    ok(ret, "expected success %lu\n", GetLastError());
     ok(len, "expected len > 0\n");
-    ok(lstrlenW( dirW ) == len - 1, "length mismatch %d != %d - 1\n", lstrlenW( dirW ), len );
+    ok(lstrlenW( dirW ) == len - 1, "length mismatch %d != %lu - 1\n", lstrlenW( dirW ), len );
     HeapFree( GetProcessHeap(), 0, dirW );
 
     CloseHandle( token );
